+++ src/ADP.Portal.Core/Git/Infrastructure/IGitOpsConfigRepository.cs
@@ -13,14 +13,16 @@ namespace ADP.Portal.Core.Git.Infrastructure
 
         Task<IEnumerable<KeyValuePair<string, Dictionary<object, object>>>> GetAllFilesAsync(GitRepo gitRepo, string path);
 
-        Task<Reference?> GetBranchAsync(GitRepo gitRepo, string branchName);
+        Task PushFilesToRepository(GitRepo gitRepoFluxServices, string branchName, Dictionary<string, Dictionary<object, object>> generatedFiles);
 
-        Task<Reference> CreateBranchAsync(GitRepo gitRepo, string branchName, string sha);
+        //Task<Reference?> GetBranchAsync(GitRepo gitRepo, string branchName);
 
-        Task<Reference> UpdateBranchAsync(GitRepo gitRepo, string branchName, string sha);
+        //Task<Reference> CreateBranchAsync(GitRepo gitRepo, string branchName, string sha);
 
-        Task<Commit?> CreateCommitAsync(GitRepo gitRepo, Dictionary<string, Dictionary<object, object>> generatedFiles, string message, string? branchName = null);
+        //Task<Reference> UpdateBranchAsync(GitRepo gitRepo, string branchName, string sha);
 
-        Task<bool> CreatePullRequestAsync(GitRepo gitRepo, string branchName, string message);
+        //Task<Commit?> CreateCommitAsync(GitRepo gitRepo, Dictionary<string, Dictionary<object, object>> generatedFiles, string message, string? branchName = null);
+
+        //Task<bool> CreatePullRequestAsync(GitRepo gitRepo, string branchName, string message);
     }
 }
