+++ test/ADP.Portal.Core.Tests/Git/Infrastructure/GitOpsConfigRepositoryTests.cs
@@ -1,8 +1,8 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.Text;
+﻿using System.Text;
 using ADP.Portal.Core.Git.Entities;
 using ADP.Portal.Core.Git.Infrastructure;
 using AutoFixture;
+using Microsoft.Extensions.Logging;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
 using NUnit.Framework;
@@ -26,7 +26,7 @@ namespace ADP.Portal.Core.Tests.Git.Infrastructure
             gitHubClientMock = Substitute.For<IGitHubClient>();
             serializer = new SerializerBuilder().WithNamingConvention(CamelCaseNamingConvention.Instance).Build();
             deserializer = new DeserializerBuilder().WithNamingConvention(CamelCaseNamingConvention.Instance).Build();
-            repository = new GitOpsConfigRepository(gitHubClientMock, deserializer, serializer);
+            repository = new GitOpsConfigRepository(gitHubClientMock, deserializer, serializer, Substitute.For<ILogger<GitOpsConfigRepository>>());
             fixture = new Fixture();
         }
